Some cleanup refactoring (#177)
