+++ deeptime/clustering/include/deeptime/clustering/metric.h
@@ -34,9 +34,7 @@ struct EuclideanMetric {
 };
 
 template<typename Metric, typename T>
-py::array_t<int> assign_chunk_to_centers(const np_array_nfc<T> &chunk,
-                                         const np_array_nfc<T> &centers,
-                                         int n_threads);
+py::array_t<int> assign_chunk_to_centers(const np_array_nfc<T> &chunk, const np_array_nfc<T> &centers, int n_threads);
 
 template<typename dtype>
 class Distances {
@@ -80,74 +78,12 @@ private:
 };
 
 template<typename dtype>
-std::unique_ptr<double[]> precomputeXX(const dtype *xs, std::size_t nXs, std::size_t dim) {
-    std::unique_ptr<double[]> xx(new double[nXs]);
-    auto xp = xx.get();
-
-    #pragma omp parallel for
-    for (std::size_t i = 0; i < nXs; ++i) {
-        xp[i] = std::inner_product(xs + i * dim, xs + i * dim + dim, xs + i * dim, static_cast<double>(0));
-    }
-    return xx;
-}
+std::unique_ptr<double[]> precomputeXX(const dtype *xs, std::size_t nXs, std::size_t dim);
 
 template<bool squared, typename Metric, typename dtype>
 Distances<dtype> computeDistances(const dtype *xs, std::size_t nXs,
                                   const dtype *ys, std::size_t nYs, std::size_t dim, const double *xxPrecomputed,
-                                  const double *yyPrecomputed) {
-    Distances<dtype> result(nXs, nYs, dim);
-    if constexpr (!std::is_same<Metric, EuclideanMetric>::value) {
-        dtype *outPtr = result.data();
-        if (squared) {
-            #pragma omp parallel for default(none) firstprivate(nXs, nYs, xs, ys, dim, outPtr)
-            for (std::size_t i = 0; i < nXs; ++i) {
-                for (std::size_t j = 0; j < nYs; ++j) {
-                    outPtr[i * nYs + j] = Metric::template compute_squared<dtype>(xs + i * dim, ys + j * dim, dim);
-                }
-            }
-        } else {
-            #pragma omp parallel for default(none) firstprivate(nXs, nYs, xs, ys, dim, outPtr)
-            for (std::size_t i = 0; i < nXs; ++i) {
-                for (std::size_t j = 0; j < nYs; ++j) {
-                    outPtr[i * nYs + j] = Metric::template compute<dtype>(xs + i * dim, ys + j * dim, dim);
-                }
-            }
-        }
-    } else {
-        dtype *outPtr = result.data();
-        // xxPrecomputed has shape (nXs,)
-        // yyPrecomputed has shape (nYs,)
-        std::unique_ptr<double[]> xx;
-        if (xxPrecomputed == nullptr) {
-            xx = precomputeXX(xs, nXs, dim);
-            xxPrecomputed = xx.get();
-        }
-        std::unique_ptr<double[]> yy;
-        if (yyPrecomputed == nullptr) {
-            yy = precomputeXX(ys, nYs, dim);
-            yyPrecomputed = yy.get();
-        }
-        {
-            // compute -2 * XY
-            #pragma omp parallel for collapse(2)
-            for (std::size_t i = 0; i < nXs; ++i) {
-                for (std::size_t j = 0; j < nYs; ++j) {
-                    outPtr[i * nYs + j] = std::inner_product(xs + i * dim, xs + i * dim + dim, ys + j * dim,
-                                                             static_cast<double>(0));
-                    outPtr[i * nYs + j] *= -2.;
-                    outPtr[i * nYs + j] += xxPrecomputed[i] + yyPrecomputed[j];
-                }
-            }
-        }
-        if (!squared) {
-            #pragma omp parallel for
-            for (std::size_t i = 0; i < result.size(); ++i) {
-                *(result.data() + i) = std::sqrt(*(result.data() + i));
-            }
-        }
-    }
-    return result;
-}
+                                  const double *yyPrecomputed);
 
 }
