--- deeptime/markov/tools/estimation/sparse/mle/newton/CMakeLists.txt
@@ -1,8 +1,0 @@
-project(objective_sparse_ops CXX)
-
-set(SRC objective_sparse_ops.cpp)
-pybind11_add_module(${PROJECT_NAME} ${SRC})
-target_link_libraries(${PROJECT_NAME} PUBLIC deeptime::common)
-if(OpenMP_FOUND)
-    target_link_libraries(${PROJECT_NAME} PUBLIC OpenMP::OpenMP_CXX)
-endif()
