--- deeptime/markov/tools/estimation/sparse/mle/newton/__init__.py
@@ -1,3 +1,0 @@
-r"""
-.. moduleauthor:: B.Trendelkamp-Schroer <benjamin DOT trendelkamp-schroer AT fu-berlin DOT de>
-"""
