--- deeptime/markov/tools/estimation/sparse/mle/newton/linsolve.py
@@ -1,29 +1,0 @@
-r"""
-.. moduleauthor:: B.Trendelkamp-Schroer <benjamin DOT trendelkamp-schroer AT fu-berlin DOT de>
-"""
-
-import numpy as np
-from scipy.sparse import issparse
-
-
-def mydot(A, B):
-    r"""Dot-product that can handle dense and sparse arrays
-
-    Parameters
-    ----------
-    A : numpy ndarray or scipy sparse matrix
-        The first factor
-    B : numpy ndarray or scipy sparse matrix
-        The second factor
-
-    Returns
-    C : numpy ndarray or scipy sparse matrix
-        The dot-product of A and B
-
-    """
-    if issparse(A) :
-        return A.dot(B)
-    elif issparse(B):
-        return (B.T.dot(A.T)).T
-    else:
-        return np.dot(A, B)
