+++ examples/methods/plot_sindy_rossler_attractor.py
@@ -11,7 +11,6 @@ Note that for this example we pass in the exact derivatives. In practice one can
 also pass in a numerical approximation in their place.
 """
 import matplotlib.pyplot as plt
-from mpl_toolkits.mplot3d import Axes3D
 import numpy as np
 from scipy.integrate import odeint
 from sklearn.preprocessing import PolynomialFeatures
