+++ tests/markov/msm/test_mlmsm.py
@@ -55,6 +55,23 @@ def test_estimator_params(reversible, statdist, sparse, maxiter, maxerr):
         assert_equal(msm.maxerr, maxerr)
 
 
+def test_fit_with_invalid_args():
+    estimator_without_model = TransitionCountEstimator(lagtime=1, count_mode='sliding')
+
+    with assert_raises(ValueError):
+        MaximumLikelihoodMSM().fit(estimator_without_model)
+    with assert_raises(ValueError):
+        MaximumLikelihoodMSM().fit_from_counts(estimator_without_model)
+    with assert_raises(ValueError):
+        class Bogus:
+            pass
+        MaximumLikelihoodMSM().fit(Bogus())
+    with assert_raises(ValueError):
+        class Bogus:
+            pass
+        MaximumLikelihoodMSM().fit_from_counts(Bogus())
+
+
 def test_weakly_connected_count_matrix():
     count_matrix = np.array([[10, 1, 0, 0], [0, 1, 1, 0], [0, 1, 1, 1], [0, 0, 0, 1]], dtype=np.float32)
     assert_equal(MaximumLikelihoodMSM().fit(count_matrix).fetch_model().n_connected_msms, 3,
@@ -209,19 +226,20 @@ class FF:
 def test_score_cv(double_well_msm_all, n_jobs):
     scenario, est, msm = double_well_msm_all
     est.lagtime = 10
+    data = scenario.dtraj[:5000]
 
     fit_fetch = FF(est)
 
     with assert_raises(ValueError):
-        vamp_score_cv(fit_fetch, trajs=scenario.dtraj, lagtime=10, n=5, r=1, dim=2, n_jobs=1, splitting_mode="noop")
+        vamp_score_cv(fit_fetch, trajs=data, lagtime=10, n=5, r=1, dim=2, n_jobs=1, splitting_mode="noop")
     with assert_raises(ValueError):
-        vamp_score_cv(fit_fetch, trajs=scenario.dtraj)  # uses blocksplit but no lagtime
-    s1 = vamp_score_cv(fit_fetch, trajs=scenario.dtraj, lagtime=10, n=5, r=1, dim=2, n_jobs=n_jobs).mean()
+        vamp_score_cv(fit_fetch, trajs=data)  # uses blocksplit but no lagtime
+    s1 = vamp_score_cv(fit_fetch, trajs=data, lagtime=10, n=5, r=1, dim=2, n_jobs=n_jobs).mean()
     assert 1.0 <= s1 <= 2.0
-    s2 = vamp_score_cv(fit_fetch, trajs=scenario.dtraj, lagtime=10, n=5, r=2, dim=2, n_jobs=n_jobs).mean()
+    s2 = vamp_score_cv(fit_fetch, trajs=data, lagtime=10, n=5, r=2, dim=2, n_jobs=n_jobs).mean()
     assert 1.0 <= s2 <= 2.0
-    se = vamp_score_cv(fit_fetch, trajs=scenario.dtraj, lagtime=10, n=5, r="E", dim=2, n_jobs=n_jobs).mean()
-    se_inf = vamp_score_cv(fit_fetch, trajs=scenario.dtraj, lagtime=10, n=5, r="E", dim=None, n_jobs=n_jobs).mean()
+    se = vamp_score_cv(fit_fetch, trajs=data, lagtime=10, n=5, r="E", dim=2, n_jobs=n_jobs).mean()
+    se_inf = vamp_score_cv(fit_fetch, trajs=data, lagtime=10, n=5, r="E", dim=None, n_jobs=n_jobs).mean()
 
 
 class TestMSMMinCountConnectivity(unittest.TestCase):
