+++ tests/markov/tools/analysis/test_expectations.py
@@ -80,7 +80,7 @@ def to_sparse_setting(setting, output_dim):
 @pytest.mark.parametrize("N", [20, 50, 0], ids=lambda x: f"nsteps={x}")
 def test_expected_counts(setting, sparse_mode, N):
     if sparse_mode:
-        setting = to_sparse_setting(setting, 10000)
+        setting = to_sparse_setting(setting, 500)
     T, mu, v, L, R = setting
     EC_n = expected_counts(T, mu, N)
     D_mu = diags(mu, 0)
@@ -95,7 +95,7 @@ def test_expected_counts(setting, sparse_mode, N):
 @pytest.mark.parametrize("statdist", [False, True], ids=lambda x: f"statdist={x}")
 def test_expected_counts_stationary(setting, sparse_mode, statdist):
     if sparse_mode:
-        setting = to_sparse_setting(setting, 10000)
+        setting = to_sparse_setting(setting, 500)
     T, mu, v, L, R = setting
     N = 20
     D_mu = diags(mu, 0)
